requirements.txt
cupy==10.6.0
equilib==0.0.1
fast_slic==0.4.0
imageio==2.9.0
matplotlib==3.5.1
numpy==1.22.3
Pillow==9.2.0
py360convert==0.1.0
pynvrtc==9.2
scikit_image==0.19.2
scipy==1.7.3
skimage==0.0
torch==1.11.0
torch_geometric==2.0.4
torch_scatter==2.0.9
torch_sparse==0.6.14
torchvision==0.12.0
tqdm==4.64.0
trimesh==3.12.8
typing_extensions==4.3.0
